refactor(core): split ClsModule to multiple modules
