+++ packages/core/src/lib/cls-initializers/use-cls.decorator.spec.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { Test } from '@nestjs/testing';
 import { CLS_ID } from '../cls.constants';
-import { ClsModule } from '../cls.module';
+import { ClsModule } from '../cls-module/cls.module';
 import { ClsService } from '../cls.service';
 import { UseCls } from './use-cls.decorator';
 import { ClsServiceManager } from '../cls-service-manager';
